Request util updates/fixes
Updated request util to allow an instance to globally be set to json as well as what header should always be sent.
Fixed bug where post data wasn’t properly set
Assume if json is set to true, response is should be json, and vice versa.  Removed try..catch to handle the mixture