Adding missing Github init, removed access_token
Github was not enableable because it the strategy wasn't initialized, so added that.  Stripped out access_token and expire from the OAuth model/table as we don't really need it, we won't need to make API calls with it. Expire doesn't make sense to be saved without the access token associated with it.